fix: component composition on main layer
