+++ src/main/factories/controllers/login/LoginValidationCompositeFactory.spec.ts
@@ -1,11 +1,11 @@
-import ValidationComposite from '../../../presentation/helpers/validators/ValidationComposite'
-import RequiredFieldValidation from '../../../presentation/helpers/validators/RequiredFieldValidation'
-import Validation from '../../../presentation/protocols/Validation'
-import EmailValidation from '../../../presentation/helpers/validators/EmailValidation'
-import EmailValidator from '../../../presentation/protocols/EmailValidator'
+import ValidationComposite from '../../../../presentation/helpers/validators/ValidationComposite'
+import RequiredFieldValidation from '../../../../presentation/helpers/validators/RequiredFieldValidation'
+import Validation from '../../../../presentation/protocols/Validation'
+import EmailValidation from '../../../../presentation/helpers/validators/EmailValidation'
+import EmailValidator from '../../../../presentation/protocols/EmailValidator'
 import LoginValidationCompositeFactory from './LoginValidationCompositeFactory'
 
-jest.mock('../../../presentation/helpers/validators/ValidationComposite')
+jest.mock('../../../../presentation/helpers/validators/ValidationComposite')
 
 class EmailValidatorStub implements EmailValidator {
   isValid (_: string): boolean {
