+++ src/main/factories/controllers/signup/SignUpValidationCompositeFactory.ts
@@ -1,8 +1,8 @@
-import Validation from '../../../presentation/protocols/Validation'
-import ValidationComposite from '../../../presentation/helpers/validators/ValidationComposite'
-import RequiredFieldValidation from '../../../presentation/helpers/validators/RequiredFieldValidation'
-import EmailValidation from '../../../presentation/helpers/validators/EmailValidation'
-import EmailValidatorAdapter from '../../adapters/EmailValidatorAdapter'
+import Validation from '../../../../presentation/protocols/Validation'
+import ValidationComposite from '../../../../presentation/helpers/validators/ValidationComposite'
+import RequiredFieldValidation from '../../../../presentation/helpers/validators/RequiredFieldValidation'
+import EmailValidation from '../../../../presentation/helpers/validators/EmailValidation'
+import EmailValidatorAdapter from '../../../adapters/EmailValidatorAdapter'
 
 const SignUpValidationCompositeFactory = {
   makeSignUpValidationComposite (): Validation {
