+++ src/main/routes/AuthenticationRoutes.ts
@@ -1,7 +1,7 @@
 import { Router } from 'express'
-import SignUpControllerFactory from '../factories/signup/SignUpControllerFactory'
+import SignUpControllerFactory from '../factories/controllers/signup/SignUpControllerFactory'
 import ExpressRouteAdapter from '../adapters/ExpressRouteAdapter'
-import LoginControllerFactory from '../factories/login/LoginControllerFactory'
+import LoginControllerFactory from '../factories/controllers/login/LoginControllerFactory'
 
 const authenticationRoutes = (router: Router): void => {
   router.post('/signup', ExpressRouteAdapter(SignUpControllerFactory.makeSignUpController()))
